улучшает окрытие и закрытие формы для загрузки фото
