+++ js/main.js
@@ -2,4 +2,4 @@ import './util.js';
 import './data.js';
 import './picture.js';
 import './create-big-picture.js';
-import './form-validation.js';
+import './open-close-download-image-popup.js';
